+++ computer_database/src/main/java/com/excilys/computer_database/persistence/dao/ComputerDAO.java
@@ -31,7 +31,7 @@ public interface ComputerDAO {
      * @param id the id of the computer
      * @return the retrieved computer itself
      */
-    Computer get(int id);
+    Computer get(long id);
 
     /**
      * Returns a computer according to the name passed.
@@ -58,14 +58,14 @@ public interface ComputerDAO {
      * @param id id of the computer you want to delete.
      * @throws SQLException thrown in case of SQL issues
      */
-    void delete(int id) throws SQLException;
+    void delete(long id) throws SQLException;
 
     /**
      * Delete all computers sharing a company that has this id.
      * @param id id of the company
      * @throws SQLException thrown for roll back purpose
      */
-    void deleteByCompany(int id) throws SQLException;
+    void deleteByCompany(long id) throws SQLException;
 
     /**
      * Retrieve the number if computers available on the db.
